requirements.txt
streamlit==1.32.0
tensorflow==2.16.1
xgboost==2.0.3
numpy==1.26.4
Pillow==10.2.0